This is ex2. the changes are:
in App.java -> first I used an algorithm to define the error
triangle(the impossible)
then I corrected a mistakes in the enum types of each statement.
Then in the function isEquilateral, isIsosceles and isTriangle there was
some mistakes such as a==a
I fixed all of it
In AppTest.java I tested each enum type to confirm the expected result.
the Error expected result is the impossible to test it.